% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cdev.generic}
\alias{cdev.generic}
\title{Condition-number-based deviation of X and Y,
for generally given X and Y.}
\usage{
cdev.generic(X, Y, always.small = TRUE)
}
\arguments{
\item{X}{m x n matrix}

\item{Y}{m x n matrix}

\item{always.small}{[=TRUE] if set to TRUE, cdev will be calculate for the smaller version of transformation matrix}
}
\description{
Much faster with thin matrix (nrow >> ncol).
}
