% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cdev}
\alias{cdev}
\title{Condition-number-based deviation of X and Y,
assuming X can be transformed into Y by scaling each columns with
a coefficient.}
\usage{
cdev(X, Y, scale = FALSE, generic = FALSE, ...)
}
\arguments{
\item{X}{a matrix}

\item{Y}{another matrix with the same shape as X}

\item{generic}{whether to use the generic definition of cdev, without assuming only column-scaling is involved between X and Y}
}
\description{
Condition-number-based deviation of X and Y,
assuming X can be transformed into Y by scaling each columns with
a coefficient.
}
