% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{community.gce}
\alias{community.gce}
\title{Community detection by GCE}
\usage{
community.gce(graph, seed, ncores = 1, version = 1)
}
\description{
Starting from a set of seeds, a vertex in the shell is added to the community if it increases community quality metric M
}
