% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{crossval}
\alias{crossval}
\title{generic function for cross-validation}
\usage{
crossval(X, ref.idx, diff.func, k = 5)
}
\arguments{
\item{X}{read count matrix}

\item{ref.idx}{the indices of features to use as references}

\item{diff.func}{the function that calculate the differences between 2 normalized matrices}

\item{k}{[=5]}
}
\description{
generic function for cross-validation
}
