% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cv.lognorm}
\alias{cv.lognorm}
\title{Coefficient of variation, unbiased estimate for log-normal variable}
\usage{
cv.lognorm(x, pseudocount = 1)
}
\description{
Coefficient of variation, unbiased estimate for log-normal variable
}
