% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fast.condNumber}
\alias{fast.condNumber}
\title{Condition number using fast.svd}
\usage{
fast.condNumber(X)
}
\description{
When X is rectangular, the SVD of XX' (or X'X) will be faster if ncol > nrow (or nrow > ncol)
Since we only care about the ratio between first and last singular values,
there's no need to calculate V'
}
