% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get.references.apcluster}
\alias{get.references.apcluster}
\title{get.references.apcluster}
\usage{
get.references.apcluster(
  m,
  cor.method = "pearson",
  min.size = 10,
  min.corr = 0.75,
  min.count = 0,
  med.quantile = 0.5,
  log.base = 2,
  debug = FALSE,
  verbose.output = FALSE,
  ...
)
}
\arguments{
\item{m}{expression matrix, genes x samples (which will be transposed internally)}

\item{cor.method}{[="pearson"] the correlation method used to construct edges between genes. Acceptable values are same as those of R base function `cor()`.}

\item{min.size}{minimum size of the candidate refernece clusters}

\item{min.count}{minimum count of a gene across all samples, for it to be retained in constructing the graph}

\item{med.quantile}{the quantile above which a gene should be expressed in all samples, for it to be retained in constructing the graph}

\item{log.base}{the base to log-transform the expression matrix}

\item{debug}{whether to print additional information when running}

\item{verbose.output}{whether to all clusters (in addition to the reference one) in the output}
}
\description{
Graph-based identify the set of references using affinity propagation as the 
underlying clustering method
}
