% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get.references.blocks}
\alias{get.references.blocks}
\title{Identify the best set of references}
\usage{
get.references.blocks(
  m,
  cor.method = "pearson",
  n.runs = 3,
  min.size = 10,
  min.corr = 0.75,
  min.count = 0,
  log.base = 2,
  out.file = "",
  debug = FALSE
)
}
\description{
Identify the best set of references
}
