% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get.references.dbscan}
\alias{get.references.dbscan}
\title{get.references.dbscan}
\usage{
get.references.dbscan(
  m,
  cor.method = "pearson",
  min.size = 3,
  min.count = 0,
  med.quantile = 0.5,
  log.base = 2,
  eps = 0.1,
  debug = FALSE,
  ...
)
}
\description{
Graph-based identify the set of references using DBSCAN as the underlying clustering method
}
