% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get.references.hclust}
\alias{get.references.hclust}
\title{get.references.hclust}
\usage{
get.references.hclust(
  m,
  cor.method = "pearson",
  min.corr = 0.75,
  min.size = 3,
  min.count = 0,
  med.quantile = 0.5,
  log.base = 2,
  debug = FALSE
)
}
\description{
get.references.hclust
}
