% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linspace}
\alias{linspace}
\title{linspace}
\usage{
linspace(min, max, n)
}
\arguments{
\item{min}{(inclusive) start of the range}

\item{max}{(inclusive) end of the range}

\item{n}{number of points, including the two ends}
}
\description{
linspace
}
