% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{measure.M}
\alias{measure.M}
\title{calculate M measure, defined as}
\usage{
measure.M(graph, community)
}
\description{
M(C) = (vol(C) - cut(C)) / (2 * cut(C))
graph    an igraph object
community    a vector of vertex indices
}
