% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{normalize.by.deges}
\alias{normalize.by.deges}
\title{Normalize by DEGES}
\usage{
normalize.by.deges(
  X,
  group,
  norm.method = "tmm",
  test.method = "edger",
  iteration = 1
)
}
\description{
Normalize by DEGES
}
