% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{normalize.by.deseq}
\alias{normalize.by.deseq}
\title{Call estimateSizeFactorsForMatrix by DESeq2}
\usage{
normalize.by.deseq(X, ...)
}
\arguments{
\item{X}{read count matrix with samples in rows and genes in columns}
}
\description{
Call estimateSizeFactorsForMatrix by DESeq2
}
