% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{normalize.by.refs}
\alias{normalize.by.refs}
\title{Normalizing (global-scaling) using references}
\usage{
normalize.by.refs(X, ref.idx, scale = TRUE)
}
\arguments{
\item{X}{a read-count matrix of the form genes x samples}

\item{ref.idx}{an integer vector specifying the column indices of X to be used as reference}
}
\description{
Normalize a read count matrix given the set of reference genes identified by id
}
