% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{normalize.by.scaleFactors}
\alias{normalize.by.scaleFactors}
\title{normalize.by.scaleFactors}
\usage{
normalize.by.scaleFactors(X, scaleFactors)
}
\arguments{
\item{X}{count matrix in the form genes x samples}
}
\description{
normalize.by.scaleFactors
}
