% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rank1.residuals}
\alias{rank1.residuals}
\title{the magnitude of the singular values except the first one}
\usage{
rank1.residuals(m)
}
\description{
the magnitude of the singular values except the first one
}
