% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{rmsd}
\alias{rmsd}
\title{Least root mean square distance between two set of points in d-dimension}
\usage{
rmsd(P, Q, transform = FALSE)
}
\description{
Find and perform the transformation (translation + rotation) on Q that minimizes the distance between 2 matrices P and Q.
Return the root mean squared of the differences
}
