% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate.counts}
\alias{simulate.counts}
\title{simulate.counts}
\usage{
\method{simulate}{counts}(
  ngenes,
  nlibs1,
  nlibs2,
  percentDE = 0.5,
  expected.lib.size = rep(1.1e+07, nlibs),
  fc = 2,
  random.seed = NULL,
  use.invChisq = TRUE,
  symmetric = FALSE,
  partiallyAsymm = TRUE,
  completelyAsymm = FALSE
)
}
\arguments{
\item{random.seed}{[NULL]}

\item{use.invChisq}{[TRUE] Use inverse chi-square or log-normal dispersion}

\item{symmetric}{Symmetric conditions have the same number of up-regulated genes}
}
\value{
an object with `counts` and `differential`. `counts` is the samples x genes matrix
}
\description{
simulate.counts
}
