% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variance.refs}
\alias{variance.refs}
\title{variance.refs}
\usage{
variance.refs(X, idx, norm = TRUE)
}
\arguments{
\item{X}{an expression matrix of the form samples x genes}

\item{idx}{indices of the reference set}
}
\description{
Calculate the average variance of the references in a given expression matrix
}
